feat(Dialogs): add new mmSimpleDialogs
