+++ Makefile.in
@@ -136,6 +136,7 @@ MMEX_OBJECTS =  \
 	mmex_mmOptionMiscSettings.o \
 	mmex_mmOptionNetSettings.o \
 	mmex_mmOptionViewSettings.o \
+	mmex_mmSimpleDialogs.o \
 	mmex_mmcheckingpanel.o \
 	mmex_mmex.o \
 	mmex_mmframe.o \
@@ -542,6 +543,9 @@ mmex_mmOptionNetSettings.o: $(srcdir)/src/mmOptionNetSettings.cpp
 mmex_mmOptionViewSettings.o: $(srcdir)/src/mmOptionViewSettings.cpp
 	$(CXXC) -c -o $@ $(MMEX_CXXFLAGS) $(srcdir)/src/mmOptionViewSettings.cpp
 
+mmex_mmSimpleDialogs.o: $(srcdir)/src/mmSimpleDialogs.cpp
+	$(CXXC) -c -o $@ $(MMEX_CXXFLAGS) $(srcdir)/src/mmSimpleDialogs.cpp
+
 mmex_mmcheckingpanel.o: $(srcdir)/src/mmcheckingpanel.cpp
 	$(CXXC) -c -o $@ $(MMEX_CXXFLAGS) $(srcdir)/src/mmcheckingpanel.cpp
 
@@ -616,7 +620,7 @@ mmex_wizard_newdb.o: $(srcdir)/src/wizard_newdb.cpp
 
 mmex_wizard_update.o: $(srcdir)/src/wizard_update.cpp
 	$(CXXC) -c -o $@ $(MMEX_CXXFLAGS) $(srcdir)/src/wizard_update.cpp
-	
+
 mmex_budget.o: $(srcdir)/src/reports/budget.cpp
 	$(CXXC) -c -o $@ $(MMEX_CXXFLAGS) $(srcdir)/src/reports/budget.cpp
