+++ src/assetdialog.cpp
@@ -15,15 +15,19 @@
  ********************************************************/
 
 #include "assetdialog.h"
-#include "mmtextctrl.h"
 #include "attachmentdialog.h"
-#include "paths.h"
 #include "constants.h"
-#include "validators.h"
+#include "mmSimpleDialogs.h"
+#include "mmtextctrl.h"
+#include "paths.h"
 #include "util.h"
+#include "validators.h"
+
 #include "model/Model_Asset.h"
 #include "model/Model_Attachment.h"
+
 #include "../resources/attachment.xpm"
+
 #include <wx/valnum.h>
 
 wxIMPLEMENT_DYNAMIC_CLASS(mmAssetDialog, wxDialog);
@@ -228,7 +232,7 @@ void mmAssetDialog::OnOk(wxCommandEvent& /*event*/)
 {
     const wxString name = m_assetName->GetValue().Trim();
     if (name.empty()){
-        mmMessageNameInvalid(m_assetName);
+        mmErrorDialogs::InvalidName(m_assetName);
         return;
     }
