+++ src/splittransactionsdialog.cpp
@@ -19,10 +19,13 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "splittransactionsdialog.h"
 #include "splitdetailsdialog.h"
 #include "constants.h"
+#include "mmSimpleDialogs.h"
 #include "util.h"
-#include <wx/statline.h>
-#include "model/Model_Category.h"
+
 #include "model/Model_Account.h"
+#include "model/Model_Category.h"
+
+#include <wx/statline.h>
 
 /*!
  * SplitTransactionDialog type definition
@@ -205,7 +208,7 @@ void SplitTransactionDialog::OnOk( wxCommandEvent& /*event*/ )
         total += entry.SPLITTRANSAMOUNT;
     if (total < 0)
     {
-        mmShowErrorMessage(this, _("Invalid Total Amount"), _("Error"));
+        mmErrorDialogs::MessageError(this, _("Invalid Total Amount"), _("Error"));
     }
     else
     {
