+++ src/transdialog.cpp
@@ -18,21 +18,25 @@
  ********************************************************/
 
 #include "transdialog.h"
-#include "mmtextctrl.h"
-#include "util.h"
-#include "splittransactionsdialog.h"
-#include "constants.h"
-#include "paths.h"
-#include "categdialog.h"
-#include "validators.h"
 #include "attachmentdialog.h"
+#include "categdialog.h"
+#include "constants.h"
+#include "mmSimpleDialogs.h"
+#include "mmtextctrl.h"
+#include "paths.h"
+#include "splittransactionsdialog.h"
+#include "util.h"
+#include "validators.h"
 #include "webapp.h"
+
 #include "model/Model_Account.h"
+#include "model/Model_Attachment.h"
 #include "model/Model_Category.h"
 #include "model/Model_Subcategory.h"
-#include "model/Model_Attachment.h"
-#include <wx/numformatter.h>
+
 #include "../resources/attachment.xpm"
+
+#include <wx/numformatter.h>
 
 wxIMPLEMENT_DYNAMIC_CLASS(mmTransDialog, wxDialog);
 
@@ -500,7 +504,7 @@ bool mmTransDialog::validateData()
     Model_Account::Data* account = Model_Account::instance().get(cbAccount_->GetValue());
     if (!account || Model_Account::type(account) == Model_Account::INVESTMENT)
     {
-        mmMessageAccountInvalid((wxWindow*)cbAccount_);
+        mmErrorDialogs::InvalidAccount((wxWindow*)cbAccount_);
         return false;
     }
     m_trx_data.ACCOUNTID = account->ACCOUNTID;
@@ -510,7 +514,7 @@ bool mmTransDialog::validateData()
         wxString payee_name = cbPayee_->GetValue();
         if (payee_name.IsEmpty())
         {
-            mmMessagePayeeInvalid((wxWindow*)cbPayee_);
+            mmErrorDialogs::InvalidPayee((wxWindow*)cbPayee_);
             return false;
         }
 
@@ -550,7 +554,7 @@ bool mmTransDialog::validateData()
         Model_Account::Data *to_account = Model_Account::instance().get(cbPayee_->GetValue());
         if (!to_account || to_account->ACCOUNTID == m_trx_data.ACCOUNTID || Model_Account::type(to_account) == Model_Account::INVESTMENT)
         {
-            mmMessageAccountInvalid((wxWindow*)cbPayee_, true);
+            mmErrorDialogs::InvalidAccount((wxWindow*)cbPayee_, true);
             return false;
         }
         m_trx_data.TOACCOUNTID = to_account->ACCOUNTID;
@@ -567,7 +571,7 @@ bool mmTransDialog::validateData()
     {
         if (local_splits.empty())
         {
-            mmMessageCategoryInvalid((wxWindow*)bCategory_);
+            mmErrorDialogs::InvalidCategory((wxWindow*)bCategory_);
             return false;
         }
     }
@@ -577,7 +581,7 @@ bool mmTransDialog::validateData()
         Model_Subcategory::Data *subcategory = Model_Subcategory::instance().get(m_trx_data.SUBCATEGID);
         if (!category || !(subcategory || m_trx_data.SUBCATEGID < 0))
         {
-            mmMessageCategoryInvalid((wxWindow*)bCategory_);
+            mmErrorDialogs::InvalidCategory((wxWindow*)bCategory_);
             return false;
         }
     }
